separated api.js from SearchPeopleResults and moved API key and URL to .env
